change dir for processor
