--- n0core/compute/kvm/__init__.py
@@ -1,1 +1,0 @@
-from .vm import VM  # NOQA
