--- n0core/lib/n0mq/__init__.py
@@ -1,1 +1,0 @@
-from .n0mq import N0MQ # noqa
