--- n0core/lib/proto/README.md
@@ -1,4 +1,0 @@
-n0stack Protobuf Schemas (Compiled files for Python)
-===
-
-For usage and details, please refer [n0stack Protobuf Schemas](../../../proto/README.md)
