--- n0core/lib/proto/__init__.py
@@ -1,25 +1,0 @@
-from .attach_network_interface_request_pb2 import AttachNetworkInterfaceRequest  # NOQA
-from .attach_volume_request_pb2 import AttachVolumeRequest  # NOQA
-from .create_ipv4_subnet_request_pb2 import CreateIPv4SubnetRequest  # NOQA
-from .create_network_request_pb2 import CreateNetworkRequest  # NOQA
-from .create_vm_request_pb2 import CreateVMRequest  # NOQA
-from .create_vm_snapshot_request_pb2 import CreateVMSnapshotRequest  # NOQA
-from .create_volume_request_pb2 import CreateVolumeRequest  # NOQA
-from .delete_ipv4_subnet_request_pb2 import DeleteIPv4SubnetRequest  # NOQA
-from .delete_network_request_pb2 import DeleteNetworkRequest  # NOQA
-from .delete_vm_request_pb2 import DeleteVMRequest  # NOQA
-from .delete_vm_snapshot_request_pb2 import DeleteVMSnapshotRequest  # NOQA
-from .delete_volume_request_pb2 import DeleteVolumeRequest  # NOQA
-from .detach_network_interface_request_pb2 import DetachNetworkInterfaceRequest  # NOQA
-from .detach_volume_request_pb2 import DetachVolumeRequest  # NOQA
-from .enums.network_type_pb2 import NetworkType  # NOQA
-from .enums.vm_power_state_pb2 import VMPowerState  # NOQA
-from .n0stack_message_pb2 import N0stackMessage  # NOQA
-from .notification_pb2 import Notification  # NOQA
-from .request_pb2 import Request  # NOQA
-from .update_ipv4_subnet_request_pb2 import UpdateIPv4SubnetRequest  # NOQA
-from .update_network_interface_request_pb2 import UpdateNetworkInterfaceRequest  # NOQA
-from .update_network_request_pb2 import UpdateNetworkRequest  # NOQA
-from .update_vm_power_state_request_pb2 import UpdateVMPowerStateRequest  # NOQA
-from .update_vm_request_pb2 import UpdateVMRequest  # NOQA
-from .vlan_network_setting_pb2 import VlanNetworkSetting  # NOQA
