--- n0core/lib/target.py
@@ -1,13 +1,0 @@
-from typing import Tuple
-
-from n0core.lib.model import Model  # NOQA
-
-
-class Target(object):
-    def __init__(self):
-        # type: () -> None
-        pass
-
-    def apply(self, model):
-        # type: (Model) -> Tuple[bool, str]
-        pass
