+++ n0core/message/notify.py
@@ -1,12 +1,12 @@
 from enum import Enum
 
-from n0core.lib.message import Message
-from n0core.lib.object import Object # NOQA
+from n0core.message import Message
+from n0core.object import Object # NOQA
 
 
 class Notify(Message):
     """
-    DO NOT REUSE.
+    DO NOT REUSE INSTANCE.
     """
 
     EVENTS = Enum("EVENTS", ["SCHEDULED", "APPLIED"])
