+++ n0core/processor/agent.py
@@ -1,7 +1,7 @@
-from n0core.lib.processor import Processor
-from n0core.lib.processor import IncompatibleMessage
-from n0core.lib.message import Message
-from n0core.lib.message.notify import Notify
+from n0core.processor import Processor
+from n0core.processor import IncompatibleMessage
+from n0core.message import Message
+from n0core.message.notify import Notify
 
 
 class Agent(Processor):
